2008-04-07  Hidehisa Akiyama  <akky@users.sourceforge.jp>
	* qt4/disp_holder.cpp:
	* qt4/disp_holder.h:
	* qt4/log_player.cpp:
	* qt4/log_player.h:
	* qt4/log_player_tool_bar.cpp:
	* qt4/log_player_tool_bar.h:
	* qt4/main_data.cpp:
	* qt4/main_data.h:
	* qt4/main_window.cpp:
	* qt4/main_window.h:
	* qt4/player_painter.cpp:
	* qt4/remote_monitor.cpp:
	* rcsslogplayer/parser.cpp:
	* rcsslogplayer/util.cpp:
	* tool/rcg3to4.cpp:
	- fixed bugs in the older version log parsing.
	- implemented a function of data segment recording.



git-svn-id: file:///home/akiyama/sserver/svn/rcss/trunk/rcsslogplayer@247 1639d2a8-b03b-4162-8b0b-03f2cde17881

+++ rcsslogplayer/util.cpp
@@ -267,7 +267,7 @@ void
 convert( const pos_t & from,
          PlayerT & to )
 {
-    Side side = static_cast< Side >( ntohs( from.side ) );
+    Side side = static_cast< Side >( static_cast< Int16 >( ntohs( from.side ) ) );
 
     to.state_ = static_cast< Int32 >( ntohs( from.enable ) );
     to.side_ = ( side == LEFT ? 'l'
